Copy over action from optimade-python-tools
Copying over the files directly related to the action from its
repository of origin: Materials-Consortia/optimade-python-tools.
Furthermore, the .gitignore is cleared and README.md updated with the
text related to the action from the original repository.
CI tests for the validator have been copied over as well, but will not
work at the moment, since starting an OPTiMaDe server in docker remains
in the optimade-python-tools repository and should not be moved here.
Instead a work-around should be found.