Update tests with new base URL runs
Make sure the validator always runs at least twice. First for the
unversioned base URL and second for the major-versioned base URL.

Use `setup_file` instead of `setup` to only run the function once for
every test file, instead of once for every single test.

Create a copy of `entrypoint.sh` in the tests folder that is used for
tests, it is an exact copy with the exception of `set -e`, which has
been commented out.