Add `create_outputs` and `results` output
The new output will be a JSON object of JSON objects.
The outer objects' keys represent the base URL, it can either be:
- `"unversioned"`
- `"vMAJOR"`
- `"vMAJOR.MINOR"`
- `"vMAJOR.MINOR.PATCH"`
where MAJOR, MINOR, and PATCH represent the current OPTIMADE API
version.

New tests to make sure the `-j/--json` option is used when
`create_outputs` is `True`.

New CI tests to make sure the output `results` is correctly created,
with the correct content depending on the base URLs checked.