Update entrypoint and Dockerfile for testing
Use virtualenv to install optimade into, removing the virtualenv at the
end.

Tried using pipx first, but this did not function as intended - so I
defaulted to virtualenv.

Updating the base of the Dockerfile to be a materials-consortia-specific
bats image, which is to be created manually prior to building the
Dockerfile image. Information about this must be added to the README.